+++ saleor/account/emails.py
@@ -1,6 +1,7 @@
 from urllib.parse import urlencode
 
 from django.conf import settings
+from django.contrib.auth.tokens import default_token_generator
 from django.urls import reverse
 from templated_email import send_templated_mail
 
@@ -8,6 +9,12 @@ from ..account import events as account_events
 from ..celeryconf import app
 from ..core.emails import get_email_base_context
 from ..core.utils import build_absolute_uri
+
+
+def send_user_password_reset_email(redirect_url, user):
+    """Trigger sending a password reset email for the given user."""
+    token = default_token_generator.make_token(user)
+    send_password_reset_email_with_url.delay(redirect_url, user.email, token, user.pk)
 
 
 def _send_password_reset_email(reset_url, recipient_email, user_id):
@@ -23,14 +30,14 @@ def _send_password_reset_email(reset_url, recipient_email, user_id):
 
 
 @app.task
-def send_password_reset_email(context, recipient, user_id):
+def send_password_reset_email(context, recipient_email, user_id):
     reset_url = build_absolute_uri(
         reverse(
             "account:reset-password-confirm",
             kwargs={"uidb64": context["uid"], "token": context["token"]},
         )
     )
-    _send_password_reset_email(reset_url, recipient, user_id)
+    _send_password_reset_email(reset_url, recipient_email, user_id)
 
 
 @app.task
